fix: code style issues

+++ pkg/file/ls.go
@@ -51,7 +51,7 @@ func (f *File) ListFiles(files []string) ([]Entry, error) {
 			Name:             meta.Name,
 			ContentType:      meta.ContentType,
 			Size:             strconv.FormatUint(meta.Size, 10),
-			BlockSize:        strconv.FormatInt(int64(uint64(meta.BlockSize)), 10),
+			BlockSize:        strconv.FormatInt(int64(meta.BlockSize), 10),
 			CreationTime:     strconv.FormatInt(meta.CreationTime, 10),
 			AccessTime:       strconv.FormatInt(meta.AccessTime, 10),
 			ModificationTime: strconv.FormatInt(meta.ModificationTime, 10),
